Update the new install service
The League Info table is not starting with a primary key of 0 - add
a primary key column to account for this; update the admin
controller to change the name from conference to league; update the
new install service to assign the right points to each pick type.
Note that pick type 4 is broken and add TODO to reflect that.