Update the submit-picks template
Update the picks controller to remove the datetime method if picks
can be submitted.  This will be updated later when pick submissions
are actually working.  Update the submit-pick template with all of
the correct lists to iterate over in the Chameleon template.

Note: this is still broken with a sqlalchemy timeout somehwere.