Add a terrible hack to make MySQL work
Add max_overflow=-1 to the dbsession file when starting MySQL to avoid
the sqlalchemy timeout problems I'm having.  This at least confirms
the app is working but I need to figure out where to add
session.close() statements to fix this.

Update the picks_controller to match the picks in the template