Add session.close() to playerpicks and remove pool from MySQL
Still trying to avoid the sqlalchemy timeouts - try a session.close()
after the picks are submitted to close the database connection.

Remove poolclass=NullPool from the dbsession which was a hack to get
around the errors.  Still have a lot of research to do on how to
return a list from the db in Pyramid AND close the database without
losing the object in memory