Update the view picks template page
Change the division team standings rank to 5 (from 4) - 5 teams in
baseball division (4 in football).  Update the category names from
football to baseball.  Add twins_wins as the tiebreaker category to
display correctly.

Note:  This page is broken as it's iterating over the list incorrectly
and showing all results 3x, not once