Add first pass of change-picks template page
Add a stub for change-picks based on submit picks.  Add a Bootstrap
alert to let the player kow they can only change up to 14 picks.
This needs to be significantly updated and include elements of
the view picks page to their original pick in one column and the
dropdown to change their pick in another column