Finally got the API call working to get the first game of the season
It helps if you remember to add ".json" to the URL.  Get the first game
season info including the teams playing, date and time.  Show it in
the admin panel.

Need to write a method to display how much time is remaining to make
your picks and add it to the submit-picks page