Refactor new season service
Refactor newseason_service to use the MSF API to get the date of
the first game (rather than user input).  Also get the time of the
game and the teams playing.  This will be used to display to the user
how much time is remaining for them to enter their picks.