Add time check for submit-picks
Use Pendulum and the database to check if the current time is less
than the time when picks are due.  If it is, let the user submit picks
or redirect them to the too-late page.  Also check if the user has
submitted picks and redirect to change-picks and vice versa