Refactor the Standings pages
The Standings homepage now displays a list of all seasons played
(instead of the standings points for the current seasons).  Add page
for each season played as a GET based on which season the user
chooses from the bullet list of all seasons played.  Add
corresponding viewmodel.  Update the your_picks viewmodel to remove
import modules not used.